// src/app/api/generated/model/SendMoneyContext.ts
/**
 * Wari teller API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface SendMoneyContext {
    securityTokenTransaction?: boolean;

    tokenValidationDesc?: string;

    wariPassOTP?: string;

    wariPassToken?: string;

    destinationCountryId?: number;

    sender?: models.Customer;

    receiver?: models.Customer;

    principal?: number;

    motif?: models.Motif;

    receptionMode?: SendMoneyContext.ReceptionModeEnum;

    paymentMean?: SendMoneyContext.PaymentMeanEnum;

}
export namespace SendMoneyContext {
    export enum ReceptionModeEnum {
        CASH = <any> 'CASH',
        WALLET = <any> 'WALLET',
        IBAN = <any> 'IBAN',
        CARD = <any> 'CARD',
        ACCOUNT = <any> 'ACCOUNT'
    }
    export enum PaymentMeanEnum {
        CASH = <any> 'CASH',
        CARD = <any> 'CARD',
        WARIPASS = <any> 'WARI_PASS',
        VOUCHER = <any> 'VOUCHER',
        WALLET = <any> 'WALLET',
        ACCOUNT = <any> 'ACCOUNT'
    }
}